mars2mars: Separate common functionality from the pure mapping rules
